+++ src/apis/http/tx.js
@@ -5,6 +5,22 @@ import apisHelper from '../helpers/apisHelper';
 const tx = {
   /**
    *
+   * @param {String} order
+   * @param {Number} limit
+   * @returns {AxiosPromise}
+   */
+  requestTxs(order, limit) {
+    return axios({
+      method: 'get',
+      baseURL: `${APIS.HTTP.BACKEND_TXS}`,
+      params: {
+        order: order || undefined,
+        limit: limit || undefined,
+      },
+    });
+  },
+  /**
+   *
    * @param {Object} parameters
    * @param {Object} pagination
    * @returns {AxiosPromise}