+++ src/views/home/HomeViewGridListItem.vue
@@ -41,7 +41,7 @@
                 />
                 <router-link
                   class="pl-1 text-decoration-none text-overline"
-                  v-text="hash"
+                  v-text="transaction.hash"
                   :to="txLink"
                 />
               </div>
@@ -54,8 +54,8 @@
 </template>
 
 <script>
-import { ROUTES } from '@/constants';
-import { dateHandler, txHandler } from '@/utils';
+import { ROUTES, TRANSACTIONS } from '@/constants';
+import { dateHandler } from '@/utils';
 
 export default {
   name: 'HomeViewGridListItem',
@@ -70,14 +70,11 @@ export default {
     date() {
       return new Date(this.transaction.timestamp).toLocaleString();
     },
-    hash() {
-      return this.transaction.txhash;
-    },
     statusClass() {
-      return !this.transaction.code ? 'info--text' : 'error--text';
+      return this.transaction.result ? 'info--text' : 'error--text';
     },
     statusText() {
-      return !this.transaction.code
+      return this.transaction.result
         ? this.$t('msgs.success')
         : this.$t('msgs.failed');
     },
@@ -87,17 +84,17 @@ export default {
         new Date()
       );
     },
-    type() {
-      return txHandler.getType_current(
-        this.transaction.tx.body.messages,
-        this.$t('labels.multiTypes')
-      );
-    },
     txLink() {
       return {
         name: ROUTES.NAME.TRANSACTIONS_DETAIL,
-        params: { id: this.hash },
+        params: { id: this.transaction.hash },
       };
+    },
+    type() {
+      if (this.transaction.msgs_number > 1) {
+        return this.$t(`labels.${TRANSACTIONS.MULTI_TYPE}`);
+      }
+      return this.transaction.types[0];
     },
   },
 };