+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.5.1] - 2023-01-26
+
+### Added
+
+- Add template to support `MsgWithdrawValidatorCommission` tx message
+
+### Changed
+
+- Switch to v2 of the backend APIs
+
 ## [4.5.0] - 2022-12-20
 
 ### Added
@@ -571,6 +581,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add RFC cross chain navigation
 - Add a switch to selcet a dark theme from navigation drawer
 
+[4.5.1]: https://github.com/commercionetwork/almerico/compare/v4.5.0...v4.5.1
 [4.5.0]: https://github.com/commercionetwork/almerico/compare/v4.4.1...v4.5.0
 [4.4.1]: https://github.com/commercionetwork/almerico/compare/v4.4.0...v4.4.1
 [4.4.0]: https://github.com/commercionetwork/almerico/compare/v4.3.5...v4.4.0